# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketSync API
  description: |
    Consolidated personal finance manager API for the Nigerian market.

    **Production base URL:** `https://pocketsync.onrender.com`

    ## Authentication
    PocketSync uses **HttpOnly cookies** — not Bearer tokens. After `POST /api/v1/auth/login`,
    the server sets `accessToken` and `refreshToken` cookies. All protected routes require
    the browser/client to send cookies (`credentials: 'include'` in fetch, `withCredentials: true` in axios).

    ## Sign-up & password reset (email OTP)
    1. `POST /api/v1/auth/register` — creates account, sends 6-digit email OTP
    2. `POST /api/v1/auth/verify-otp` with `purpose: signup` — marks email verified
    3. `POST /api/v1/auth/login` — blocked until email is verified (`403`)
    4. Forgot password: `POST /api/v1/auth/forgot-password` → `POST /api/v1/auth/reset-password`

    OTP resend cooldown is **30 seconds**. Codes expire in **10 minutes**.
    Without `RESEND_API_KEY`, OTPs are logged to the server console; in development the code
    may also appear in the JSON response as `devOtp`.

    ## BVN onboarding (mocked — Phase 2)
    After email verification and login, complete onboarding:
    1. `GET /api/v1/onboarding/status` — route user to the correct step
    2. `POST /api/v1/onboarding/bvn/submit` — mock BVN + phone validation, sends SMS OTP
    3. `POST /api/v1/onboarding/bvn/verify-otp` — verifies phone, discovers mock accounts
...

...

    ## Amounts
...

...
servers:
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Uses `refreshToken` cookie (path-scoped to this endpoint).
        Rotates both tokens on success.
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed — new cookies set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token refreshed
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly cookie set by POST /api/v1/auth/login
...
  responses:
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          description: Present in development only
...